+++ source/subtitle_sync/init.d/install_ffs.sh
@@ -10,11 +10,3 @@ if [ ! "$ffspy" ]; then
 else
     echo "**** ffsubsync already installed ****"
 fi
-
-piso639=$(python3 -m pip list | grep python-iso639)
-if [ ! "$piso639" ]; then
-    echo "**** Installing python-iso639 ****"
-    python3 -m pip install -U python-iso639
-else
-    echo "**** python-iso639 already installed ****"
-fi